move helper function location
