+++ tests/transforms/test_experimental/test_transform_program.py
@@ -28,7 +28,6 @@ from pennylane.transforms.core import (
 from pennylane.transforms.core.transform_program import (
     _apply_postprocessing_stack,
     _batch_postprocessing,
-    _prune_dynamic_transform,
     null_postprocessing,
 )
 from pennylane.typing import Result, ResultBatch
@@ -107,27 +106,6 @@ class TestUtilityHelpers:
 
         out2 = _apply_postprocessing_stack(results, [postprocessing2, postprocessing1])
         assert out2 == (4.0, 9.0)
-
-    def test_prune_dynamic_transform(self):
-        """Tests that prune dynamic transform works."""
-
-        program1 = TransformProgram(
-            [
-                qml.transforms.dynamic_one_shot,
-                qml.transforms.sum_expand,
-                qml.transforms.dynamic_one_shot,
-            ]
-        )
-        program2 = TransformProgram(
-            [
-                qml.transforms.dynamic_one_shot,
-                qml.transforms.sum_expand,
-            ]
-        )
-
-        _prune_dynamic_transform(program1, program2)
-        assert len(program1) == 1
-        assert len(program2) == 2
 
 
 class TestTransformProgramDunders:
